+++ src/plopp/__init__.py
@@ -15,7 +15,7 @@ plt.ioff()
 
 from .graph import show_graph
 from .model import Node, node, input_node, widget_node
-from .wrappers import plot, figure, slicer
+from .functions import figure, plot, slicer
 
 from . import data
