+++ src/plopp/functions/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
+
+# flake8: noqa E402, F401
+
+from .figure import figure
+from .plot import plot
+from .slicer import slicer
