+++ src/plopp/functions/common.py
@@ -1,0 +1,11 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
+
+from matplotlib import get_backend
+
+
+def is_interactive_backend():
+    """
+    Return `True` if the current backend used by Matplotlib is the widget backend.
+    """
+    return 'ipympl' in get_backend()
